Merged PR 15831: Add initial Loop structure
The Loop structure will be the basis for iteration going forward. Each Loop structure is created with a LoopConditionTunnel, which defines a mutable Boolean variable that is used to control when the Loop terminates. The Loop also supports a simple value-moving Tunnel and a LoopBorrowTunnel, which behaves similarly to the FlatSequence BorrowTunnel; both this and the LoopConditionTunnel are paired with a LoopTerminateLifetimeTunnel.