Merged PR 18534: Allow PassthroughTerminalPair to specify whether it is related by lifetime to any output parameters
DetermineVariablesTransform will initially treat a passthrough input and output as separate variables if the passthrough parameter is related by lifetime to any outputs; it is up to the node's implementation of SetVariableTypesTransform to merge the input and output variables when appropriate.

Currently this only applies to SelectReferenceNode, but will also be used for various vector/slice nodes.